# contracts/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/screens": {
      "post": {
        "operationId": "ScreensController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateScreenDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScreenDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a screen profile.",
        "tags": [
          "screens"
        ]
      },
      "get": {
        "operationId": "ScreensController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "items",
                    "meta"
                  ],
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ScreenDto"
                      }
                    },
                    "meta": {
                      "$ref": "#/components/schemas/PageMetaDto"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List screens (filter by location/status/orientation/use/tag/group).",
        "tags": [
          "screens"
        ]
      }
    },
...
    "/screens/{id}/monitoring": {
      "get": {
        "operationId": "ScreenMonitoringController_screenMonitoring",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScreenMonitoringDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Live status + telemetry + cache/sync + latest screenshot for a screen.",
        "tags": [
          "screens"
        ]
      }
    },
    "/screens/{id}/screenshots": {
      "get": {
        "operationId": "ScreenMonitoringController_screenshotHistory",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "items",
                    "meta"
                  ],
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ScreenshotSummaryDto"
                      }
                    },
                    "meta": {
                      "$ref": "#/components/schemas/PageMetaDto"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Paginated screenshot history for a screen, newest first.",
        "tags": [
          "screens"
        ]
      }
    },
    "/screens/{id}/commands": {
      "get": {
        "operationId": "ScreenMonitoringController_listCommands",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "items",
                    "meta"
                  ],
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/DeviceCommandDto"
                      }
                    },
                    "meta": {
                      "$ref": "#/components/schemas/PageMetaDto"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Paginated remote-command history for a screen.",
        "tags": [
          "screens"
        ]
      },
      "post": {
        "operationId": "ScreenMonitoringController_issueCommand",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IssueCommandDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceCommandDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Issue a remote command to a screen.",
        "tags": [
          "screens"
        ]
      }
    },
    "/screens/{id}/commands/{commandId}": {
      "get": {
        "operationId": "ScreenMonitoringController_getCommand",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "commandId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceCommandDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get one remote command, including its result or error.",
        "tags": [
          "screens"
        ]
      }
    },
    "/screens/{id}/actions/force-sync": {
      "post": {
        "operationId": "ScreenMonitoringController_forceSync",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceCommandDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Force a full re-sync of the screen’s assets.",
        "tags": [
          "screens"
        ]
      }
    },
    "/screens/{id}/actions/refresh-manifest": {
      "post": {
        "operationId": "ScreenMonitoringController_refreshManifest",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceCommandDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Ask the screen to re-fetch its playback manifest.",
        "tags": [
          "screens"
        ]
      }
    },
    "/screens/{id}/actions/restart-playback": {
      "post": {
        "operationId": "ScreenMonitoringController_restartPlayback",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceCommandDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Restart playback on the screen.",
        "tags": [
          "screens"
        ]
      }
    },
    "/screens/{id}/actions/clear-cache": {
      "post": {
        "operationId": "ScreenMonitoringController_clearCache",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceCommandDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Clear the screen’s local asset cache.",
        "tags": [
          "screens"
        ]
      }
    },
    "/screens/{id}/actions/take-screenshot": {
      "post": {
        "operationId": "ScreenMonitoringController_takeScreenshot",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceCommandDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Ask the screen to capture and upload a screenshot.",
        "tags": [
          "screens"
        ]
      }
    },
    "/screens/{id}/actions/reboot": {
      "post": {
        "operationId": "ScreenMonitoringController_reboot",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceCommandDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Reboot the device.",
        "tags": [
          "screens"
        ]
      }
    },
...
  },
  "info": {
    "title": "Wizer Signage API",
    "description": "Wizer Signage — multi-tenant digital signage SaaS platform REST API.",
    "version": "0.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "PageMetaDto": {
        "type": "object",
        "properties": {
          "page": {
            "type": "number",
            "example": 1
          },
          "pageSize": {
            "type": "number",
            "example": 25
          },
          "total": {
            "type": "number",
            "example": 137,
            "description": "Total matching rows, ignoring pagination."
          },
          "totalPages": {
            "type": "number",
            "example": 6
          }
        },
        "required": [
          "page",
          "pageSize",
          "total",
          "totalPages"
        ]
      },
...
      "CreateScreenDto": {
        "type": "object",
        "properties": {}
      },
      "NestedRefDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name"
        ]
      },
      "TagDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "companyId": {
            "type": "string",
            "description": "Owning tenant. Derived from the caller's token, never the body."
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "name": {
            "type": "string"
          },
          "color": {
            "type": "object",
            "nullable": true,
            "example": "#2F80ED"
          },
          "description": {
            "type": "object",
            "nullable": true
          },
          "type": {
            "type": "string",
            "enum": [
              "CONTENT",
              "SCREEN",
              "BOTH"
            ],
            "description": "Which resources the tag may be applied to."
          }
        },
        "required": [
          "id",
          "companyId",
          "createdAt",
          "updatedAt",
          "name"
        ]
      },
      "ScreenDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "companyId": {
            "type": "string",
            "description": "Owning tenant. Derived from the caller's token, never the body."
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "name": {
            "type": "string"
          },
          "code": {
            "type": "object",
            "nullable": true
          },
          "description": {
            "type": "object",
            "nullable": true
          },
          "locationId": {
            "type": "object",
            "nullable": true
          },
          "location": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/NestedRefDto"
              }
            ]
          },
          "deviceIdentifier": {
            "type": "object",
            "nullable": true,
            "description": "Hardware id reported by the paired player; null until pairing."
          },
          "pairedAt": {
            "type": "object",
            "nullable": true,
            "format": "date-time"
          },
          "lastHeartbeatAt": {
            "type": "object",
            "nullable": true,
            "format": "date-time"
          },
          "lastSyncAt": {
            "type": "object",
            "nullable": true,
            "format": "date-time"
          },
          "appVersion": {
            "type": "object",
            "nullable": true
          },
          "audioEnabled": {
            "type": "boolean"
          },
          "volume": {
            "type": "number",
            "minimum": 0,
            "maximum": 100
          },
          "muted": {
            "type": "boolean"
          },
          "muteSchedule": {
            "type": "object",
            "nullable": true
          },
          "workingHours": {
            "type": "object",
            "nullable": true
          },
          "kioskEnabled": {
            "type": "boolean"
          },
          "hasKioskPin": {
            "type": "boolean",
            "description": "Whether an exit PIN is set. The PIN itself is stored hashed and is never returned — this boolean replaces it."
          },
          "autoStartEnabled": {
            "type": "boolean"
          },
          "powerControlMode": {
            "type": "object",
            "nullable": true
          },
          "fallbackContentId": {
            "type": "object",
            "nullable": true
          },
          "currentContentId": {
            "type": "object",
            "nullable": true
          },
          "currentPlaylistId": {
            "type": "object",
            "nullable": true
          },
          "storageUsedBytes": {
            "type": "string",
            "nullable": true,
            "description": "Bytes, as a STRING. The column is a 64-bit integer and would lose precision as a JSON number, so PrismaService serialises BigInt via toJSON."
          },
          "storageTotalBytes": {
            "type": "string",
            "nullable": true,
            "description": "Bytes, as a string."
          },
          "capabilities": {
            "type": "object",
            "nullable": true
          },
          "heartbeatIntervalSeconds": {
            "type": "number"
          },
          "latitude": {
            "type": "number",
            "nullable": true
          },
          "longitude": {
            "type": "number",
            "nullable": true
          },
          "notes": {
            "type": "object",
            "nullable": true
          },
          "tags": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TagDto"
            }
          },
          "groups": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NestedRefDto"
            }
          }
        },
        "required": [
          "id",
          "companyId",
          "createdAt",
          "updatedAt",
          "name",
          "audioEnabled",
          "volume",
          "muted",
          "kioskEnabled",
          "hasKioskPin",
          "autoStartEnabled",
          "heartbeatIntervalSeconds",
          "tags",
          "groups"
        ]
      },
...
      "ScreenTelemetryDto": {
        "type": "object",
        "properties": {
          "playbackState": {
            "type": "object",
            "nullable": true
          },
          "networkStatus": {
            "type": "object",
            "nullable": true
          },
          "currentContentId": {
            "type": "object",
            "nullable": true
          },
          "currentPlaylistId": {
            "type": "object",
            "nullable": true
          },
          "currentScheduleId": {
            "type": "object",
            "nullable": true
          },
          "manifestVersion": {
            "type": "object",
            "nullable": true
          },
          "uptimeSeconds": {
            "type": "object",
            "nullable": true
          },
          "appVersion": {
            "type": "object",
            "nullable": true
          },
          "deviceModel": {
            "type": "object",
            "nullable": true,
            "description": "Device `modelName`, renamed by the view."
          },
          "osVersion": {
            "type": "object",
            "nullable": true
          },
          "platform": {
            "type": "object",
            "nullable": true
          }
        }
      },
      "ScreenCacheDto": {
        "type": "object",
        "properties": {
          "syncStatus": {
            "type": "object",
            "nullable": true
          },
          "manifestSource": {
            "type": "object",
            "nullable": true
          },
          "cacheSizeBytes": {
            "type": "object",
            "nullable": true,
            "description": "Serialised as a string — a 64-bit column, not a JSON number."
          },
          "availableStorageBytes": {
            "type": "object",
            "nullable": true,
            "description": "Serialised as a string — a 64-bit column, not a JSON number."
          },
          "requiredAssets": {
            "type": "number"
          },
          "cachedAssets": {
            "type": "number"
          },
          "failedDownloads": {
            "type": "number"
          },
          "lastSyncAt": {
            "type": "object",
            "nullable": true,
            "format": "date-time"
          },
          "lastError": {
            "type": "object",
            "nullable": true,
            "description": "Device `lastSyncError`, renamed by view."
          }
        },
        "required": [
          "requiredAssets",
          "cachedAssets",
          "failedDownloads"
        ]
      },
      "LatestScreenshotDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "takenAt": {
            "type": "string",
            "format": "date-time"
          },
          "type": {
            "type": "string",
            "enum": [
              "AUTO",
              "MANUAL"
            ]
          },
          "url": {
            "type": "object",
            "nullable": true,
            "description": "Short-lived signed URL."
          }
        },
        "required": [
          "id",
          "takenAt",
          "type"
        ]
      },
...
      "ScreenMonitoringDto": {
        "type": "object",
        "properties": {
          "screenId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "ONLINE",
              "OFFLINE",
              "WARNING",
              "UNPAIRED",
              "PAIRING",
              "DISABLED",
              "ARCHIVED"
            ],
            "description": "DERIVED on every read from the last heartbeat, not a stored column — a paired screen with no fresh heartbeat reads OFFLINE without any scheduler having written that."
          },
          "warningReason": {
            "type": "object",
            "nullable": true,
            "description": "Set only when status is WARNING."
          },
          "lastHeartbeatAt": {
            "type": "object",
            "nullable": true,
            "format": "date-time"
          },
          "heartbeatIntervalSeconds": {
            "type": "object",
            "nullable": true
          },
          "paired": {
            "type": "boolean",
            "description": "True only for a device row in ACTIVE status."
          },
          "telemetry": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/ScreenTelemetryDto"
              }
            ]
          },
          "cache": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/ScreenCacheDto"
              }
            ]
          },
          "latestScreenshot": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/LatestScreenshotDto"
              }
            ]
          },
          "capabilities": {
            "$ref": "#/components/schemas/ScreenCapabilitiesDto"
          }
        },
        "required": [
          "screenId",
          "name",
          "status",
          "paired",
          "capabilities"
        ]
      },
      "ScreenshotSummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "takenAt": {
            "type": "string",
            "format": "date-time"
          },
          "type": {
            "type": "string",
            "enum": [
              "AUTO",
              "MANUAL"
            ]
          },
          "mimeType": {
            "type": "object",
            "nullable": true
          },
          "fileSizeBytes": {
            "type": "object",
            "nullable": true,
            "description": "Serialised as a string — a 64-bit column, not a JSON number."
          },
          "url": {
            "type": "object",
            "nullable": true,
            "description": "Short-lived signed URL. null when signing failed; the caller re-reads rather than caching it."
          }
        },
        "required": [
          "id",
          "takenAt",
          "type"
        ]
      },
      "DeviceCommandDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "screenId": {
            "type": "string"
          },
          "commandType": {
            "type": "string",
            "enum": [
              "FORCE_SYNC",
              "REFRESH_MANIFEST",
              "RESTART_PLAYBACK",
              "CLEAR_CACHE",
              "TAKE_SCREENSHOT",
              "REBOOT_DEVICE",
              "UNPAIR_DEVICE",
              "RELOAD_CONFIG"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "DELIVERED",
              "RUNNING",
              "SUCCEEDED",
              "FAILED",
              "EXPIRED",
              "CANCELLED"
            ]
          },
          "payload": {
            "type": "object",
            "additionalProperties": true,
            "nullable": true,
            "description": "Command-specific arguments; `{}` for the parameterless actions."
          },
          "result": {
            "type": "object",
            "additionalProperties": true,
            "nullable": true,
            "description": "Whatever the device reported back; null until it completes."
          },
          "error": {
            "type": "object",
            "nullable": true
          },
          "issuedById": {
            "type": "object",
            "nullable": true,
            "description": "null for a system-issued command."
          },
          "deliveredAt": {
            "type": "object",
            "nullable": true,
            "format": "date-time"
          },
          "startedAt": {
            "type": "object",
            "nullable": true,
            "format": "date-time"
          },
          "completedAt": {
            "type": "object",
            "nullable": true,
            "format": "date-time"
          },
          "expiresAt": {
            "type": "object",
            "nullable": true,
            "format": "date-time",
            "description": "A command not collected by this time is marked EXPIRED on the next poll."
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "screenId",
          "commandType",
          "status",
          "createdAt",
          "updatedAt"
        ]
      },
      "IssueCommandDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}